Trying out links to unstructured LSQ stencil plots
This Euler solver uses an "extended stencil" which includes all cells which share a vertex with the "green cell".  The "green cell" is the cell colored in green here.  For every cell we need to reconstruct some differentiable quantities.  Since the mesh is, in general, unstructured (we take no advantage of the sample mesh's particular ordered triangular structure) the program needs some way to reconstruct flow properties over the domain.  The least squared method is one way to do it!  We pre-compute the stencil weights which ensure that the averaged quantities at the cell centers are indeed the least squares fit to the surrounding data in the stencil.  Here we have pictured one cell, in green, and its surrounding stencil -- i.e. the surrounding cells from which it draws data.